// src/types/TypeOfAddress.ts
export type TypeOfAddress = {
  buildingName?: string;
  country?: string;
  deliveryAddress?: string;
  level?: string;
  lotSection?: string;
  postcode?: string;
  state?: string;
  streetName?: string;
  streetNumber?: string;
  streetSuffix?: string;
  streetType?: string;
  suburb?: string;
  unit?: string;
};

export default TypeOfAddress;